+++ include/clap/private/macros.h
@@ -25,20 +25,26 @@
 #   endif
 #endif
 
-#if defined(__cplusplus) && __cplusplus >= 201103L
+#if defined(_MSVC_LANG)
+#   define CLAP_CPLUSPLUS _MSVC_LANG
+#elif defined(__cplusplus)
+#   define CLAP_CPLUSPLUS __cplusplus
+#endif
+
+#if defined(CLAP_CPLUSPLUS) && CLAP_CPLUSPLUS >= 201103L
 #   define CLAP_HAS_CXX11
 #   define CLAP_CONSTEXPR constexpr
 #else
 #   define CLAP_CONSTEXPR
 #endif
 
-#if defined(__cplusplus) && __cplusplus >= 201703L
+#if defined(CLAP_CPLUSPLUS) && CLAP_CPLUSPLUS >= 201703L
 #   define CLAP_HAS_CXX17
 #   define CLAP_NODISCARD [[nodiscard]]
 #else
 #   define CLAP_NODISCARD
 #endif
 
-#if defined(__cplusplus) && __cplusplus >= 202002L
+#if defined(CLAP_CPLUSPLUS) && CLAP_CPLUSPLUS >= 202002L
 #   define CLAP_HAS_CXX20
 #endif